Add Impreza 2020/2021 support (reduced torque)
I choose to keep these as two separate CARs, as there is the
possibility that they will need different tuning values. For
the moment I've put both as the tune from @letsdudiss1. Both
these cars are torque limited.

We have an fpv2 for the Impreza 2020, but only fpv1 for the 2021.