Update substitute.yaml to support Subaru Forester 2022 - 2023
Ran into a bug where my 2022 Subaru Forester failed to match the substitute value when checking torque params in interfaces.py. It fingerprinted with the candidate "SUBARU FORESTER 2022 - 2023" but since substitute.py only has "SUBARU FORESTER 2022" it failed to map. This fixed the issue and allowed openpilot to correctly map to the torque params.